Fixed the issue for constraints: Only one borrow request must  be approved for multiple requests in a same book. Approving multiple borrow requests with the same book will result a message output saying that the book is still being borrowed by someone.
If a book is still being borrowed by someone, the book won't be displayed in the available books.

However, multiple borrow requests can still be made with the same book before a request regarding with the book is approved. After that, the book cannot be requested to be borrowed anymore until the current borrower returns the book.